backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
groq==0.4.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pgvector
pydantic==2.6.1
python-dotenv==1.0.0
sentence-transformers==2.5.1
